refactor drawer-item component to overlay-item and overlay-item-content component
